% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnr.R
\docType{package}
\name{rnr}
\alias{rnr}
\alias{rnr-package}
\title{rnr: A package for computing Rosenbaum and Rubin sensitivity}
\description{
The \code{rnr} package provides functions for computing sensitivity of
counterfactual estimates under assumptions of unobserved confounding.
}
