% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitize.R
\name{sensitize}
\alias{sensitize}
\title{Compute the sensitivity-adjusted estimates of predicted outcome given
treatment/control}
\usage{
sensitize(df, q, dp, d0, d1, debug = FALSE)
}
\arguments{
\item{df}{data frame to analyze; must include columns $id: column used to
identify each row (for future debugging) $resp: Observed (binary) response,
e.g., FTA $treat: Observed (binary) treatment, e.g., bail_set $resp_ctl:
Predicted probability of positive resp given control, e.g., E[FTA(bail=0) |
bail = 0] $resp_trt: Predicted probability of positive resp given
treatment, e.g., E[FTA(bail=1) | bail = 1] $p_trt: predicted probability of
setting bail for observed x}

\item{q}{p(u = 1 | x)}

\item{dp}{change in log-odds of treat = 1 if u = 1}

\item{d0}{change in log-odds of resp = 1 if treat = 0 and u = 1}

\item{d1}{change in log-odds of resp = 1 if treat = 1 and u = 1}

\item{debug}{logical, whether or not to return columns of intermediate
variables for debuging purposes}
}
\value{
A data frame with the columns resp_ctl and resp_trt updated according
  to the sensitivity parameters.
  If debug = TRUE, returned data frame will also contain columns of
  intermediate variables computed for sensitivity, starting with "v_rnr_"
  (e.g., v_rnr_gamma), with the original response estimates renamed as
     v_rnr_resp_trt_trt = resp_trt
     v_rnr_resp_ctl_ctl = resp_ctl
}
\description{
Compute the sensitivity-adjusted estimates of predicted outcome given
treatment/control
}
\note{
If debug = FALSE (default), all columns that start with "v_rnr_" will
  be removed from the final result
}
\examples{
d <- data.frame(id = 1, resp = 1, treat = 0,
                resp_ctl = .2, resp_trt = .3, p_trt = .5)
sensitize(d)
}
