% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitize.R
\name{sensitize}
\alias{sensitize}
\title{Generic sensitizing for Rosenbaum & Rubin sensitivity analysis}
\usage{
sensitize(d, q, dp, d0, d1, debug)
}
\arguments{
\item{d}{data to sensitize}

\item{q}{p(u = 1 | x)}

\item{dp}{change in log-odds of treat = 1 if u = 1}

\item{d0}{change in log-odds of response = 1 if treat = 0 and u = 1}

\item{d1}{change in log-odds of response = 1 if treat = 1 and u = 1}

\item{debug}{logical, whether or not to return columns of intermediate
variables for debugging purposes}
}
\value{
a sensitized data structure
}
\description{
Generic sensitizing for Rosenbaum & Rubin sensitivity analysis
}
