% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitize.R
\name{sensitize}
\alias{sensitize}
\title{Generic sensitizing for Rosenbaum & Rubin sensitivity analysis}
\usage{
sensitize(obj, q, dp, d0, d1, ...)
}
\arguments{
\item{obj}{data to sensitize}

\item{q}{p(u = 1 | x)}

\item{dp}{change in log-odds of treat = 1 if u = 1}

\item{d0}{change in log-odds of response = 1 if treat = 0 and u = 1}

\item{d1}{change in log-odds of response = 1 if treat = 1 and u = 1}

\item{...}{additional arguments required to sensitize object}
}
\value{
a sensitized object, identical to, or inheriting the class of
  original \code{obj}
}
\description{
Generic sensitizing for Rosenbaum & Rubin sensitivity analysis
}
